@@ -2,19 +2,20 @@
   "name": "client",
   "version": "0.1.0",
   "private": true,
+  "homepage": "./client/build",
   "dependencies": {
-    "@types/jest": "^29.2.0",
     "fetch-intercept": "^2.4.0",
-    "https-proxy-agent": "^7.0.2",
-    "react": "0.0.0-experimental-b36ae8d7a-20231207",
-    "react-dom": "0.0.0-experimental-b36ae8d7a-20231207",
+    "node-sass": "^8.0.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-router-dom": "^6.19.0"
   },
   "scripts": {
-    "start": "cross-env NODE_ENV=development webpack serve --config config/webpack.dev.js",
-    "build": "cross-env NODE_ENV=production webpack --config config/webpack.prod.js",
-    "lint": "eslint . src config || true",
-    "test": "jest --coverage"
+    "start": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
+    "watch": "sass --watch ./src/styles/index.scss:./public/index.css",
+    "compile-sass": "sass ./src/styles/index.scss ./public/index.css"
   },
   "browserslist": {
     "production": [
@@ -35,38 +36,15 @@
     "@babel/plugin-transform-runtime": "^7.12.15",
     "@babel/preset-env": "^7.19.4",
     "@babel/preset-react": "^7.18.6",
+    "@vitejs/plugin-react": "^4.3.1",
     "babel-jest": "^29.2.2",
     "babel-loader": "^8.1.0",
-    "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "7.0.0",
-    "cross-env": "^7.0.2",
-    "css-loader": "^5.0.0",
-    "css-minimizer-webpack-plugin": "^5.0.1",
-    "eslint": "^7.12.1",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-airbnb-base": "^14.2.0",
-    "eslint-config-prettier": "^6.15.0",
-    "eslint-import-resolver-webpack": "^0.13.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-webpack-plugin": "^2.1.0",
-    "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.0.0-alpha.7",
-    "jest": "^29.2.2",
-    "jest-environment-jsdom": "^29.2.2",
-    "mini-css-extract-plugin": "^1.0.0",
-    "node-sass": "^8.0.0",
-    "postcss-loader": "^4.0.4",
-    "postcss-preset-env": "^9.3.0",
-    "react-test-renderer": "^18.2.0",
-    "sass-loader": "^10.0.3",
-    "style-loader": "^2.0.0",
-    "terser-webpack-plugin": "^5.1.1",
-    "url-loader": "^4.1.1",
-    "webpack": "^5.1.3",
+    "vite": "^5.3.1",
+    "vite-plugin-simple-html": "^0.1.2",
+    "vite-plugin-svgr": "^4.2.0",
+    "webpack": "^5.91.0",
     "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^4.15.1",
+    "webpack-dev-server": "^4.15.2",
     "webpack-merge": "^5.2.0"
   }
 }
