@@ -1,7 +1,7 @@
-module.exports = class UserDto {
-  email;
-  id;
-  isActivated;
+export default class UserDto {
+  // email;
+  // id;
+  // isActivated;
 
   constructor(model) {
     this.email = model.email;
