@@ -1,25 +1,27 @@
-const { check, validationResult } = require('express-validator');
+import { check, validationResult } from 'express-validator';
 
-exports.registerValidator = () => [
+export const registerValidator = (req, res, next) => {
   check('email')
     .isEmail()
-    .withMessage('Not valid email'),
-  check('password')
-    .isLength({ min: 6 })
-    .withMessage('Min length of password has to be 6 symbols')
-]
+    .withMessage('Not valid email')
+    check('password')
+      .isLength({ min: 6 })
+      .withMessage('Min length of password has to be 6 symbols');
+  next();
+}
 
-exports.loginValidator = () => [
-  check('email',
-    'Please, enter a valid email'
-  ).normalizeEmail().isEmail(),
-  check(
-    'password',
-    'Min length of password has to be 6 symbols'
-  ).exists()
-]
+export const loginValidator = (req, res, next) => {
+    check('email',
+      'Please, enter a valid email'
+    ).normalizeEmail().isEmail()
+    check(
+      'password',
+      'Min length of password has to be 6 symbols'
+    ).exists();
+  next();
+};
 
-exports.resultsValidator = (req) => {
+export const resultsValidator = (req) => {
   const messages = [];
 
   if (!validationResult(req).isEmpty()) {
@@ -30,4 +32,10 @@ exports.resultsValidator = (req) => {
     }
   }
   return messages
+}
+
+export default {
+  registerValidator,
+  loginValidator,
+  resultsValidator
 }
