@@ -1,12 +1,11 @@
-const Router = require('express')
-const router = new Router()
+import Router from 'express';
+import authRouter from './auth.route';
+import linkRouter from './link.route';
+import recipeRouter from './recipe.route';
 
-const authRouter = require('./auth.route')
-const linkRouter = require('./link.route')
-const recipeRouter = require('./recipe.route')
+const router = new Router();
+router.use('/auth', authRouter);
+router.use('/links', linkRouter);
+router.use('/recipes', recipeRouter);
 
-router.use('/auth', authRouter)
-router.use('/link', linkRouter)
-router.use('/recipe', recipeRouter)
-
-module.exports = router;
+export default router;
