@@ -1,22 +1,22 @@
-const { Router } = require('express');
-const Link = require('../models/Link');
+import Router from 'express';
+import Link from '../models/Link';
+
 const router = Router();
 
 router.get('/:code', async (req, res) => {
-    try {
-        const link = await Link.findOne({ code: req.params.code });
+  try {
+    const link = await Link.findOne({ code: req.params.code });
 
-        if (link) {
-            link.clicks++;
-            await link.save();
-            return res.redirect(link.from);
-        }
-
-        res.status(404).json('Link has been not found');
-
-    } catch (e) {
-        res.status(500).json({ message: 'Something went wrong. Please, try again later.' });
+    if (link) {
+      link.clicks++;
+      await link.save();
+      return res.redirect(link.from);
     }
-})
 
-module.exports = router;
+    res.status(404).json('Link has been not found');
+  } catch (e) {
+    res.status(500).json({ message: 'Something went wrong. Please, try again later.' });
+  }
+});
+
+export default router;
