@@ -1,54 +1,52 @@
-const jwt = require('jsonwebtoken')
-const config = require('config')
-const TokenModel = require('../models/Token')
+import jwt from 'jsonwebtoken';
+import config from 'config';
+import TokenModel from '../models/Token';
 
 class TokenService {
   generateTokens(payload) {
-    const accessToken = jwt.sign(payload, config.get('accessTokenSecret'), { expiresIn: '35m' })
-    const refreshToken = jwt.sign(payload, config.get('refreshTokenSecret'), { expiresIn: '90m' })
+    const accessToken = jwt.sign(payload, config.get('accessTokenSecret'), { expiresIn: '35m' });
+    const refreshToken = jwt.sign(payload, config.get('refreshTokenSecret'), { expiresIn: '90m' });
 
-    return { accessToken, refreshToken }
+    return { accessToken, refreshToken };
   }
 
-  async saveToken (userId, refreshToken) {
-    const tokenData = await TokenModel.findOne({ user: userId })
+  async saveToken(userId, refreshToken) {
+    const tokenData = await TokenModel.findOne({ user: userId });
 
     if (tokenData) {
-      tokenData.refreshTokens.push(refreshToken)
-
-      return tokenData.save()
+      tokenData.refreshTokens.push(refreshToken);
+      return tokenData.save();
     }
-
-    return await TokenModel.create({ user: userId, refreshTokens: refreshToken })
+    return await TokenModel.create({ user: userId, refreshTokens: refreshToken });
   }
 
-  async validateAccessToken (token) {
+  async validateAccessToken(token) {
     try {
-      return jwt.verify(token, config.get('accessTokenSecret'))
+      return jwt.verify(token, config.get('accessTokenSecret'));
     } catch (e) {
-      return null
+      return null;
     }
   }
 
-  async validateRefreshToken (token) {
+  async validateRefreshToken(token) {
     try {
-      return jwt.verify(token, config.get('refreshTokenSecret'))
+      return jwt.verify(token, config.get('refreshTokenSecret'));
     } catch (e) {
-      return null
+      return null;
     }
   }
 
-  async findToken (token) {
+  async findToken(token) {
     try {
-      return await TokenModel.findOne({ refreshTokens: token })
+      return await TokenModel.findOne({ refreshTokens: token });
     } catch (e) {
-      return null
+      return null;
     }
   }
 
-  async removeToken (refreshToken) {
-    return await TokenModel.deleteOne({ refreshTokens: refreshToken })
+  async removeToken(refreshToken) {
+    return TokenModel.deleteOne({ refreshTokens: refreshToken });
   }
 }
 
-module.exports = new TokenService()
+export default new TokenService();
