@@ -1,7 +1,8 @@
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 
 const useStorage = (storageName) => {
   if (!storageName) return;
+  const [storageVal, setStorageVal] = useState(null);
 
   const getStorageValue = () => {
     const val = localStorage.getItem(storageName);
@@ -15,19 +16,26 @@ const useStorage = (storageName) => {
       }
     }
     return null;
-  }
+  };
 
-  const [storageVal, setStorageVal] = useState(getStorageValue);
+  useEffect(() => {
+    setStorageVal(() => getStorageValue(storageName));
+    window.addEventListener('storage', syncLocalStorage);
+    return () => {
+      window.removeEventListener('storage', syncLocalStorage);
+    };
+  }, []);
+
 
   const updateStorage = (val) => {
     localStorage.setItem(
       storageName,
       JSON.stringify(val));
-  }
+  };
 
   const clearStorage = () => {
     localStorage.removeItem(storageName);
-  }
+  };
 
   const syncLocalStorage = (e) => {
     if (e.key === storageName) {
@@ -35,20 +43,11 @@ const useStorage = (storageName) => {
     }
   };
 
-  useEffect(() => {
-    window.addEventListener('storage', syncLocalStorage);
-
-    return () => {
-      window.removeEventListener('storage', syncLocalStorage);
-    };
-  }, []);
-
-
   return {
     storageVal,
     updateStorage,
     clearStorage
-  }
-}
+  };
+};
 
 export default useStorage;
